Setup wayland desktop in Raspberry Pi setup script

+++ bin/sh/setup-rpi.sh
@@ -2,45 +2,52 @@
 
 set -euo pipefail
 
+cd ~
+
 echo "Hi! This wizard will help you setup your Rasberry Pi (5)."
 echo "Written by me, for me, myself and i :)"
-echo
-echo "To start off, please enter the hostname set during imaging:"
-echo "(This should be excluding any suffixes such as .local etc.)"
-read -p "Hostname: " HOSTNAME
 
-echo "Great! Now onto setting up SSH and GPG keys. If I'm not mistaken,"
-echo "there should be an USB stick with those on it. Expected ls(1) output:"
-echo
-echo "ssh priv.asc pub.asc"
-echo
-echo "Please insert your USB stick now. Press any key to continue."
-read
-echo "The drive is going to be mounted at /mnt. Output from lsblk(1):"
-echo
-lsblk
-echo
-echo "Please enter the name of the drive (ie. /dev/sda1)."
-read -p "Name: " USB
+read -p "Is this the intial setup? [y/N]: " confirm
 
-echo "==> Mounting drive @ /mnt"
-sudo mount "$USB" /mnt
+case "$confirm" in
+    [Yy])
+        echo "Great! Let's start getting your SSH and GPG keys. If I'm not mistaken,"
+        echo "there should be an USB stick with those on it. Expected ls(1) output:"
+        echo
+        echo "ssh priv.asc pub.asc"
+        echo
+        echo "Please insert your USB stick now. Press any key to continue."
+        read
+        echo "The drive is going to be mounted at /mnt. Output from lsblk(1):"
+        echo
+        lsblk
+        echo
+        echo "Please enter the name of the drive (ie. /dev/sda1)."
+        read -p "Name: " USB
 
-echo "==> Copying over SSH keys."
-[ ! -d ~/.ssh ] && mkdir ~/.ssh
-cp -r /mnt/ssh/. ~/.ssh
-sudo chown -R $(whoami) ~/.ssh
-sudo chmod 600 ~/.ssh/*
+        echo "==> Mounting drive @ /mnt"
+        sudo mount "$USB" /mnt
 
-echo "==> Copying over GPG keys (will be imported later)."
-cp /mnt/priv.asc ~/
-cp /mnt/pub.asc ~/
+        echo "==> Copying over SSH keys."
+        [ ! -d ~/.ssh ] && mkdir ~/.ssh
+        cp -r /mnt/ssh/. ~/.ssh
+        sudo chown -R $(whoami) ~/.ssh
+        sudo chmod 600 ~/.ssh/*
 
-echo "==> Unmounting drive."
-sudo umount /mnt
+        echo "==> Copying over GPG keys (will be imported later)."
+        cp /mnt/priv.asc ~/
+        cp /mnt/pub.asc ~/
 
-echo "==> Removing preinstalled bloatware."
-sudo apt remove geany thonny lxterminal
+        echo "==> Unmounting drive."
+        sudo umount /mnt
+
+        echo "==> Removing preinstalled bloatware."
+        sudo apt remove geany thonny lxterminal
+        ;;
+    *)
+        echo "Well that skips a lot of steps :)"
+        ;;
+esac
 
 echo "==> Performing full system upgrade."
 sudo apt update
@@ -53,11 +60,16 @@ sudo truncate -s 0 /etc/issue
 echo "==> Installing console fonts."
 sudo apt install terminus
 
-echo "Installed Terminus font. I recommend 14x16 for the font-size but"
-echo "it's up to you ofc. Press any key to launch console-setup."
-read
+read -p "Installed Terminus font. Launch console-setup? [y/N]: " confirm
 
-sudo dpkg-reconfigure console-setup
+case "$confirm" in
+    [Yy])
+        sudo dpkg-reconfigure console-setup
+        ;;
+    *)
+        echo "OK. You can always rerun the script to do console-setup anyway."
+        ;;
+esac
 
 echo "==> Downloading GPG keys for additional repositories."
 sudo mkdir -p /etc/apt/keyrings
@@ -77,25 +89,34 @@ sudo apt update
 echo "==> Installing base system."
 sudo apt install neovim rsync git curl htop wget gh pup gum bat eza fd-find ripgrep yt-dlp pass pass-otp imv mpv playerctl mosh aerc gpg gpg-agent bash sl cmatrix dosfstools ntfs-3g imagemagick lolcat cowsay fortune-mod keychain
 
-echo "==> Linking binaries"
+echo "==> Installing build tools."
+sudo apt install build-essential make meson checkinstall
+
+if [ ! -f /usr/bin/meson-install ];
+    sudo wget -qO /usr/bin/meson-install https://raw.githubusercontent.com/keithbowes/meson-install/refs/heads/main/meson-install
+    sudo chmod +x /usr/bin/meson-install
+fi
+
+echo "==> Linking binaries."
 [ ! -e /usr/bin/bat ] && sudo ln -s /usr/bin/batcat /usr/bin/bat
 
-echo "==> Setting up binaries in $HOME"
-mkdir -p "$HOME/bin/$HOSTNAME"
-[ ! -L "$HOME/bin/$HOSTNAME.local" ] && ln -s "$HOME/bin/$HOSTNAME" "$HOME/bin/$HOSTNAME.local"
+echo "==> Setting up binaries in $HOME."
+mkdir -p "$HOME/bin/$(hostname)"
+[ ! -L "$HOME/bin/$(hostname).local" ] && ln -s "$HOME/bin/$(hostname)" "$HOME/bin/$(hostname).local"
 
-echo "==> Installing Helix"
+echo "==> Installing Helix."
 HELIX_VERSION=25.01.1
 HELIX_ARCHIVE=helix-$HELIX_VERSION-aarch64-linux
 
+cd /tmp
 wget https://github.com/helix-editor/helix/releases/download/$HELIX_VERSION/$HELIX_ARCHIVE.tar.xz
 tar -xf $HELIX_ARCHIVE.tar.xz && rm $HELIX_ARCHIVE.tar.xz
-cp $HELIX_ARCHIVE/hx "$HOME/bin/$HOSTNAME/hx"
+cp $HELIX_ARCHIVE/hx "$HOME/bin/$(hostname)/hx"
 mkdir -p ~/.local/share/helix/runtime
 cp -r $HELIX_ARCHIVE/runtime/. ~/.local/share/helix/runtime
 rm -r $HELIX_ARCHIVE
 
-echo "==> Pulling dotfiles"
+echo "==> Pulling dotfiles."
 export GIT_SSH_COMMAND="ssh -i $HOME/.ssh/sweet -o IdentitiesOnly=yes"
 
 cd ~
@@ -153,6 +174,34 @@ echo "==> Setting up mailboxes."
 mkdir -p "$XDG_CONFIG_HOME/aerc"
 rsync -avP nov:accounts.conf "$XDG_CONFIG_HOME/aerc/"
 
+echo "==> Installing wayland desktop."
+sudo apt install sway dbus xwayland seatd alacritty tofi wob kanshi mako-notifier swaybg grim slurp wl-clipboard clipman wlsunset swayidle pipewire pipewire-pulse pipewire-bin wireplumber xdg-desktop-portal xdg-desktop-portal-wlr adwaita-icon-theme
+
+echo "==> Installing dependencies for swaylock-effects."
+sudo apt install libwayland-dev wayland-protocols scdoc libxkbcommon-dev libcairo2-dev libgdk-pixbuf-2.0-dev
+
+echo "==> Downloading swaylock-effects."
+[ -d /tmp/swaylock-effects ] && sudo rm -rf /tmp/swaylock-effects
+git clone https://github.com/mortie/swaylock-effects /tmp/swaylock-effects
+cd /tmp/swaylock-effects
+
+echo "==> Building swaylock-effects."
+meson build
+ninja -C build
+echo "Swaylock, with fancy effects" > description-pak
+# The following pkgversion number is a dummy :)
+sudo checkinstall -y --pkgname swaylock-effects --pkgversion 2000 --provides swaylock meson-install
+
+echo "==> Installing dependencies for autotiling."
+sudo apt install python3-i3ipc
+
+echo "==> Downloading autotiling."
+if [ ! -f /usr/bin/autotiling ];
+    sudo wget -qO /usr/bin/autotiling https://raw.githubusercontent.com/nwg-piotr/autotiling/refs/heads/master/main.py
+    sudo chmod +x /usr/bin/autotiling
+fi
+
 echo "==> Finalizing"
+cd ~
 git remote set-url origin du:meta/dotfiles
 source ~/.bashrc
